+++ OverchargeNG/src/app/services/login/login.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { User } from '../models/User';
+import { User } from '../../models/User';
 
 @Injectable({
   providedIn: 'root'
